+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Phormium Changelog
 ==================
 
+0.4 / 2013-07-17
+----------------
+
+* Added support for custom queries via Connection object
+* Added Model->merge()
+* Changed Model::get() to accept PK as an array
+* Added Model::find()
+* Added Model::exists()
+
 0.3 / 2013-06-14
 ----------------